requirements.txt
aiogram==3.0.0
python-dotenv==1.0.0
aiosqlite==0.19.0
reportlab==4.0.0
aiofiles==23.2.0
